+++ src/js/types.ts
@@ -1,0 +1,5 @@
+export enum CordovaDevicePlatform {
+  Ios = 'ios',
+  Android = 'android',
+  Unknown = 'unknown',
+}