+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - build(internal): Switch to eslint
 - build(android): Bump Android SDK to v4.1.0 (#187)
 - feat: Add global error handler wrapper method for Ionic (#190)
+- build(ios): Bump sentry-cocoa to 6.1.4 (#194)
+- feat: Add Native Wrapper and Cordova Transport (#194)
 
 ## v0.17.0